+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## 0.13.0
 
 + Added module `argparse` with class `LetterSubparsers()`
-+ module `sqlalchemy`: 
++ Module `sqlalchemy`: 
     * removed deprecated alias `entity_base()`. use `declarative_base()` instead.
     * fix imports.
++ Module `functools`: add `cooldown()`
 
 ## 0.12.6